+++ src/app_engine/gcm_notify.py
@@ -18,7 +18,7 @@ GCM_API_KEY_PATH = 'gcm_api_key'
 GCM_API_URL = 'https://android.googleapis.com/gcm/send'
 
 GCM_MESSAGE_TYPE_BYE = 'BYE'
-GCM_MESSAGE_TYPE_INVITE = 'INVITE'
+GCM_MESSAGE_TYPE_INVITE = constants.ACK_TYPE_INVITE
 GCM_MESSAGE_TYPE_RINGING = 'RINGING'
 GCM_MESSAGE_TYPE_KEY = 'type'
 GCM_MESSAGE_REASON_KEY = 'reason'
