+++ src/app_engine/join_page_test.py
@@ -225,124 +225,6 @@ class JoinPageHandlerTest(test_utilities.BasePageHandlerTest):
     self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
                               constants.RESPONSE_INVALID_ROOM)
 
-  def testAckRingingInvalidRoom(self):
-    self.addTestData()
-
-    # Room doesn't exist.
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'invalidCallee',
-        constants.PARAM_ROOM_ID: 'newRoom',
-    }
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_INVALID_ROOM)
-
-  def testAckRingingInvalidCallee(self):
-    self.addTestData()
-    # Callee doesn't exist.
-    room_id = 'callercallee'
-    self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
-                              constants.RESPONSE_SUCCESS)
-
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'invalidCallee',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_INVALID_CALLEE)
-
-  def testAckRingingWrongRoom(self):
-    self.addTestData()
-    room_id = 'callercallee'
-    self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
-                              constants.RESPONSE_SUCCESS)
-
-    # Callee not allowed in this room.
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee2gcm1',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_INVALID_CALLEE)
-
-  def testAckRinging(self):
-    self.addTestData()
-    room_id = 'callercallee'
-    self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
-                              constants.RESPONSE_SUCCESS)
-
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee1gcm1',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    self.clearGCMPayloads()
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_SUCCESS)
-
-    expected_payloads = [
-        self.createGCMRingingPayload(
-            ['caller1gcm1'],
-            room_id),
-    ]
-    self.verifyGCMPayloads(expected_payloads)
-
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee1gcm2',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_SUCCESS)
-    self.verifyGCMPayloads(expected_payloads)
-
-  def testAckRingingRoomFull(self):
-    self.addTestData()
-    room_id = 'callercallee'
-    self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
-                              constants.RESPONSE_SUCCESS)
-    self.requestAcceptAndVerify(room_id, 'callee1gcm1',
-                                constants.RESPONSE_SUCCESS)
-
-    # Room already full.
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee1gcm1',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_INVALID_ROOM)
-
-  def testAckRingingEmptyRoom(self):
-    self.addTestData()
-    room_id = 'callercallee1'
-    self.requestCallAndVerify(room_id, 'caller1gcm1', 'callee1',
-                              constants.RESPONSE_SUCCESS)
-    self.requestLeaveAndVerify(room_id, 'caller1gcm1',
-                               constants.RESPONSE_SUCCESS)
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee1gcm1',
-        constants.PARAM_ROOM_ID: room_id,
-    }
-
-    response = self.makePostRequest('/ack', json.dumps(body))
-    self.verifyResultCode(response, constants.RESPONSE_INVALID_ROOM)
-
-  def testAckInvalidInput(self):
-    body = {
-        constants.PARAM_TYPE: constants.TYPE_RINGING,
-        constants.PARAM_CALLEE_GCM_ID: 'callee1gcm1',
-        constants.PARAM_ROOM_ID: 'room',
-    }
-    self.checkInvalidRequestsJsonResult('/ack', body.keys())
 
 if __name__ == '__main__':
   unittest.main()
